Attempt to add Quiescence Search
Attempted to add a quiescence search to go two
more depths whenever a piece is captured. However,
the ai still seems to be performing bad and making
stupid mistakes. At first I believe there was an
error in the evaluation function, but it performs well
at first before it begins to weigh every action in its
favor. Might have to explore any possible errors in the
minimax algorithm itslef.