Fixed Evaluation Algorithm/Minimax
Found the error that was causing the ai to not work
properly. The evaulation function returned positive
scores for both white/black if the player was in favor
in the minimax Algorithm, this caused an issue when
searching through moves as it would assume the enemy
player would minimize their score even though evaulation
return positive values for advantages. Fixed by making
evaulation return positive for white and negative for black.
Then went inside minimax algorithm to maximize score if
ai was white and minimize if ai was black. This seemed to
work as the ai was able to beat me and the 250 elo bot on
chess.com, which it was unable to previously.